Separated authorization part to oauth.js
