+++ scraper/app.js
@@ -1,16 +1,14 @@
 /******************************************************************************/
 /* app.js                                                                     */
 /* Author: Seung Jae Lee                                                      */
-/* Parts of code from                                                         */
-/* https://developers.google.com/google-apps/calendar/quickstart/node         */
 /******************************************************************************/
 
 var fs = require('fs');
-var readline = require('readline');
 var google = require('googleapis');
-var googleAuth = require('google-auth-library');
 var db_module = require('./db');
+var oauth = require('./oauth');
 var scraper = require('./scraper');
+
 
 const PROJECT_MODE_ERROR = 'Error: PROJECT_MODE not set. Cannot set up database. Did you activate the virtual environment in the Django project?';
 
@@ -26,15 +24,8 @@ if (process.env.PROJECT_MODE === 'development') {
     process.exit(1);
 }
 
-// read/write access except delete for gmailAutho
-var SCOPES = ['https://www.googleapis.com/auth/gmail.modify'];
-var TOKEN_DIR = (process.env.HOME || process.env.HOMEPATH ||
-        process.env.USERPROFILE) + '/.credentials/';
-var TOKEN_PATH = TOKEN_DIR + 'gmail-nodejs-foodmap.json';
-
-
 if(process.env.PROJECT_MODE === 'production') {
-    authorize(JSON.parse(process.env.client_secret), main);
+    oauth.authorize(JSON.parse(process.env.client_secret), main);
 }
 else if(process.env.PROJECT_MODE === 'development') {
     // Load client secrets from a local file.
@@ -45,97 +36,12 @@ else if(process.env.PROJECT_MODE === 'development') {
         }
         // Authorize a client with the loaded credentials, then call the
         // main program.
-        authorize(JSON.parse(content), main);
+        oauth.authorize(JSON.parse(content), main);
     });
 }
 else {
     console.error(PROJECT_MODE_ERROR);
     process.exit(1);
-}
-
-/**
- * Create an OAuth2 client with the given credentials, and then execute the
- * given callback function.
- *
- * @param {Object} credentials The authorization client credentials.
- * @param {function} callback The callback to call with the authorized client.
- */
-function authorize(credentials, callback) {
-    var clientSecret = credentials.installed.client_secret;
-    var clientId = credentials.installed.client_id;
-    var redirectUrl = credentials.installed.redirect_uris[0];
-    var auth = new googleAuth();
-    var oauth2Client = new auth.OAuth2(clientId, clientSecret, redirectUrl);
-
-    // Use JSON given from process environment
-    if (process.env.PROJECT_MODE === 'production') {
-        oauth2Client.credentials = JSON.parse(process.env.CREDENTIALS);
-        callback(oauth2Client);
-    }
-    // Check if we have previously stored a token.
-    else if (process.env.PROJECT_MODE === 'development') {
-        fs.readFile(TOKEN_PATH, function(err, token) {
-            if (err) {
-                getNewToken(oauth2Client, callback);
-            } else {
-                oauth2Client.credentials = JSON.parse(token);
-                callback(oauth2Client);
-            }
-        });
-    }
-    else {
-        console.error(PROJECT_MODE_ERROR);
-        process.exit(1);
-    }
-}
-
-/**
- * Get and store new token after prompting for user authorization, and then
- * execute the given callback with the authorized OAuth2 client.
- *
- * @param {google.auth.OAuth2} oauth2Client The OAuth2 client to get token for.
- * @param {getEventsCallback} callback The callback to call with the authorized
- *     client.
- */
-function getNewToken(oauth2Client, callback) {
-    var authUrl = oauth2Client.generateAuthUrl({
-        access_type: 'offline',
-        scope: SCOPES
-    });
-    console.log('Authorize this app by visiting this url: ', authUrl);
-    var rl = readline.createInterface({
-        input: process.stdin,
-        output: process.stdout
-    });
-    rl.question('Enter the code from that page here: ', function(code) {
-        rl.close();
-        oauth2Client.getToken(code, function(err, token) {
-            if (err) {
-                console.log('Error while trying to retrieve access token', err);
-                return;
-            }
-            oauth2Client.credentials = token;
-            storeToken(token);
-            callback(oauth2Client);
-        });
-    });
-}
-
-/**
- * Store token to disk be used in later program executions.
- *
- * @param {Object} token The token to store to disk.
- */
-function storeToken(token) {
-    try {
-        fs.mkdirSync(TOKEN_DIR);
-    } catch (err) {
-        if (err.code != 'EEXIST') {
-            throw err;
-        }
-    }
-    fs.writeFile(TOKEN_PATH, JSON.stringify(token));
-    console.log('Token stored to ' + TOKEN_PATH);
 }
 
 /**
